+++ core/src/main/scala/org/ensime/core/RichPresentationCompiler.scala
@@ -161,6 +161,11 @@ trait RichCompilerControl extends CompilerControl with RefactoringControl with C
       new SemanticHighlighting(this).symbolDesignationsInRegion(p, tpes)
     ).getOrElse(SymbolDesignations(new File("."), List.empty))
 
+  def askImplicitInfoInRegion(p: Position): List[ImplicitInfo] =
+    askOption(
+      new ImplicitAnalyzer(this).implicitDetails(p)
+    ).getOrElse(List.empty)
+
   def askClearTypeCache(): Unit = clearTypeCache()
 
   def askNotifyWhenReady(): Unit = ask(setNotifyWhenReady)